Split QRCodeGenerator further
